# apollo-openapi.yaml
openapi: 3.0.1
info:
  title: Apollo OpenAPI
  description: |
    <p>Apollo配置中心OpenAPI接口文档</p>

    <h2>认证方式</h2>
    <p>所有 API 接口都需要通过 <code>Authorization</code> header 进行身份验证。</p>

    <h3>获取 Token 的方式：</h3>
    <ol>
      <li><b>Portal 管理界面获取</b>：登录 Portal → 管理员工具 → 开放平台授权管理 → 创建第三方应用，获取 Token。</li>
      <li><b>Token 格式</b>：<code>Authorization: token_value</code></li>
      <li><b>Token 权限</b>：按应用/环境/命名空间授予，建议不同用途分别创建。</li>
    </ol>

    <h3>使用示例</h3>
    <pre><code class="language-bash">curl -X GET "http://localhost:8070/openapi/v1/apps" \
    -H "Authorization: your_token_here"</code></pre>

  version: 1.0.0
security:
  - ApiKeyAuth: []
...
paths:
...
  /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}:
    get:
      summary: 获取指定集群信息 (original openapi)
      operationId: getCluster
      deprecated: false
      description: GET /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}
      tags:
        - Cluster Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: clusterName
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功获取集群信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenClusterDTO'
              example:
                dataChangeCreatedBy: apollo
                dataChangeLastModifiedBy: ops-user
                dataChangeCreatedTime: '2024-01-15T10:30:00.000Z'
                dataChangeLastModifiedTime: '2024-01-18T09:15:00.000Z'
                name: production
                appId: sample-app
          headers: {}
        '404':
          description: 集群不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
    delete:
      summary: 删除集群 (new added)
      operationId: deleteCluster
      deprecated: false
      description: DELETE /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}
      tags:
        - Cluster Management
      parameters:
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: clusterName
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: operator
          in: query
          description: '操作人用户名，openapi外部调用时需传入'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: '集群删除成功'
        '400':
          description: '删除失败，集群可能包含配置'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        '403':
          description: '权限不足'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        '404':
          description: '集群不存在'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
          headers: {}
  /openapi/v1/envs/{env}/apps/{appId}/clusters:
    post:
      summary: 创建集群 (original openapi)
      operationId: createCluster
      deprecated: false
      description: POST /openapi/v1/envs/{env}/apps/{appId}/clusters
      tags:
        - Cluster Management
      parameters:
        - name: appId
          in: path
          description: 应用ID
          required: true
          schema:
            type: string
        - name: env
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenClusterDTO'
        required: true
      responses:
        '200':
          description: 集群创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenClusterDTO'
              example:
                dataChangeCreatedBy: apollo
                dataChangeLastModifiedBy: apollo
                dataChangeCreatedTime: '2024-01-15T10:30:00.000Z'
                dataChangeLastModifiedTime: '2024-01-15T10:30:00.000Z'
                name: test-cluster
                appId: sample-app
          headers: {}
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Apollo OpenAPI访问令牌
        **获取方式：**
        1. 登录Apollo Portal管理界面
        2. 进入"管理员工具" -> "开放平台授权管理"
        3. 创建第三方应用，获取Token
        **使用格式：**
        ```
        Authorization: your_token_here
        ```
        **权限说明：**
        - Token具有对应应用的读写权限
        - 不同Token可能有不同的环境和命名空间权限
        - 建议为不同用途创建不同的Token
  schemas:
...
    OpenClusterDTO:
      type: object
      properties:
        dataChangeCreatedBy:
          type: string
          description: 集群创建者用户名，记录是谁创建了这个集群
        dataChangeLastModifiedBy:
          type: string
          description: 集群最后修改者用户名，记录最后一次修改集群信息的用户
        dataChangeCreatedTime:
          type: string
          description: 集群创建时间，ISO 8601格式的时间戳
          example: "2025-09-29T12:34:56Z"
        dataChangeLastModifiedTime:
          type: string
          description: 集群最后修改时间，ISO 8601格式的时间戳
          example: "2025-09-29T12:34:56Z"
        name:
          type: string
          description: 集群名称，在同一应用和环境下唯一标识一个集群
        appId:
          type: string
          description: 所属应用的唯一标识符
        id:
          type: integer
          description: 集群的唯一标识符
          format: int64
        parentClusterId:
          type: integer
          description: 父集群的ID
          format: int64
        comment:
          type: string
          description: 集群的备注说明
      description: Apollo集群信息数据传输对象，表示应用在特定环境下的集群配置
...
    ExceptionResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP 状态码
        message:
          type: string
          description: 错误信息
        timestamp:
          type: string
          format: date-time
          description: 时间戳
        exception:
          type: string
          description: 错误类名
      example:
        status: 404
        message: "Resource not found"
        timestamp: "2025-09-29T12:34:56Z"
        exception: "NotFoundException"
...
servers: []